Don't recursively scan for pyproject.toml when calculating the cache key (#260)
We don't really need to scan for `pyproject.toml` files recursively,
since we only have one in the root of the repository. This should make
the cache key calculation more efficient and less error prone, as when
using qemu, there are some files that are not accessible and the hash
calculation fails.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>